Obtains DESCRIPTION metadata from CRAN for each package.

This is a wrapper around tools::CRAN_package_db and may be deprecated in future versions of the package.

Usage

getCranDescription(
  pkg,
  repos = getOption("repos"),
  type = "source",
  pkgs = pkgDep(pkg, repos = repos, type = type)
)

Arguments

pkg

Character vector of packages.

repos

URL(s) of the 'contrib' sections of the repositories, e.g. "https://cran.us.r-project.org". Passed to available.packages()

type

Possible values are (currently) "source", "mac.binary" and "win.binary": the binary types can be listed and downloaded but not installed on other platforms. Passed to download.packages().

pkgs

Character vector of packages to download

Examples

if (interactive()) {
  getCranDescription(c("igraph", "ggplot2", "XML"),
                     repos = c(CRAN = getOption("minicran.mran"))
  )
}